LRN&L2norm review comments fixed
